requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
mccabe==0.6.1
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
